home page button

+++ webapp.py
@@ -211,9 +211,14 @@ def create_room():
 @app.route('/delete_room', methods=['GET', 'POST'])
 def delete_room():
 	if request.method == 'POST':
-	    collection_rooms.delete_one({'_id': ObjectId(request.json['room_id'])})
-	    collection_messages.delete_many({'room': request.json['room_id']})
-	    return Response(dumps({'success': True}), mimetype='application/json')
+	    if collection_rooms.count() != 1 {
+	      collection_rooms.delete_one({'_id': ObjectId(request.json['room_id'])})
+	      collection_messages.delete_many({'room': request.json['room_id']})
+	      return Response(dumps({'success': 'true'}), mimetype='application/json')
+	    }
+	    else {
+	      return Response(dumps({'success': 'false'}), mimetype='application/json')
+	    }
 
 @app.route('/create_space', methods=['GET', 'POST'])
 def create_space():
